framework/tests/test_distributed.py
```python
import unittest
from distributed import DistributedScraper

class TestDistributedScraper(unittest.TestCase):

    def setUp(self):
        self.distributed_scraper = DistributedScraper()

    def test_distribute_task(self):
        task = {"url": "https://example.com", "type": "data"}
        result = self.distributed_scraper.distribute_task(task)
        self.assertTrue(result)

    def test_get_task_status(self):
        task_id = "12345"
        status = self.distributed_scraper.get_task_status(task_id)
        self.assertIn(status, ["queued", "running", "completed", "failed"])

    def test_get_task_result(self):
        task_id = "12345"
        result = self.distributed_scraper.get_task_result(task_id)
        self.assertIsInstance(result, dict)

    def test_handle_failure(self):
        task_id = "12345"
        error = "Timeout error"
        result = self.distributed_scraper.handle_failure(task_id, error)
        self.assertTrue(result)

if __name__ == '__main__':
    unittest.main()
```